+++ frontend/test/frontend_tests/runner.cljs
@@ -28,6 +28,7 @@
    [frontend-tests.logic.pasting-in-containers-test]
    [frontend-tests.main-errors-test]
    [frontend-tests.plugins.context-shapes-test]
+   [frontend-tests.plugins.format-test]
    [frontend-tests.plugins.interactions-test]
    [frontend-tests.plugins.page-active-validation-test]
    [frontend-tests.plugins.page-test]
@@ -85,6 +86,7 @@
     frontend-tests.plugins.context-shapes-test
     frontend-tests.plugins.page-active-validation-test
     frontend-tests.plugins.interactions-test
+    frontend-tests.plugins.format-test
     frontend-tests.plugins.page-test
     frontend-tests.plugins.parser-test
     frontend-tests.plugins.tokens-test